Revert Changes To attributes.go
