+++ cmd/query/app/apiv3/grpc_handler.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
 	"github.com/jaegertracing/jaeger/cmd/query/app/querysvc/v2/querysvc"
-	"github.com/jaegertracing/jaeger/internal/jpcommon"
 	"github.com/jaegertracing/jaeger/internal/jptrace"
 	"github.com/jaegertracing/jaeger/internal/proto/api_v3"
 	"github.com/jaegertracing/jaeger/internal/storage/v2/api/tracestore"
@@ -75,7 +74,7 @@ func (h *Handler) internalFindTraces(
 		TraceQueryParams: tracestore.TraceQueryParams{
 			ServiceName:   query.GetServiceName(),
 			OperationName: query.GetOperationName(),
-			Attributes:    jpcommon.PlainMapToPcommonMap(query.GetAttributes()),
+			Attributes:    jptrace.PlainMapToPcommonMap(query.GetAttributes()),
 			SearchDepth:   int(query.GetSearchDepth()),
 		},
 		RawTraces: query.GetRawTraces(),
