+++ internal/jptrace/attributes_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package jpcommon
+package jptrace
 
 import (
 	"testing"
